Add SeqTransducer that wraps Transform
