+++ xnmt/__init__.py
@@ -30,7 +30,6 @@ import xnmt.eval_task
 import xnmt.evaluator
 import xnmt.exp_global
 import xnmt.experiment
-import xnmt.ff
 import xnmt.fixed_size_att
 import xnmt.hyper_parameters
 import xnmt.inference
